Extract DynamoDBQueryTrackerCreator
